+++ src/tt/state/title.rs
@@ -179,7 +179,7 @@ impl State for TitleState {
                 gl::Viewport(
                     0,
                     0,
-                    (p_width as f32 / 4. * (3. + rcr)) as i32,
+                    (p_width / 4. * (3. + rcr as f64)) as i32,
                     p_height as i32,
                 );
                 gl::Enable(gl::GL_CULL_FACE);
@@ -220,13 +220,13 @@ impl State for TitleState {
             gl::MatrixMode(gl::GL_PROJECTION);
             gl::LoadIdentity();
             let ratio_threshold = 480. / 640.;
-            let screen_ratio = p_height as f32 / p_width as f32;
+            let screen_ratio = p_height / p_width;
             if screen_ratio >= ratio_threshold {
                 gl::Frustum(
                     -screen.near_plane() as f64,
                     screen.near_plane() as f64,
-                    (-screen.near_plane() * screen_ratio) as f64,
-                    (screen.near_plane() * screen_ratio) as f64,
+                    -screen.near_plane() as f64 * screen_ratio,
+                    screen.near_plane() as f64 * screen_ratio,
                     0.1,
                     screen.far_plane() as f64,
                 );
@@ -234,10 +234,10 @@ impl State for TitleState {
                 // This allows to see at least what can be seen horizontally and vertically
                 // with the default ratio -- arnodb
                 gl::Frustum(
-                    (-screen.near_plane() * ratio_threshold / screen_ratio) as f64,
-                    (screen.near_plane() * ratio_threshold / screen_ratio) as f64,
-                    (-screen.near_plane() * ratio_threshold) as f64,
-                    (screen.near_plane() * ratio_threshold) as f64,
+                    -screen.near_plane() as f64 * ratio_threshold / screen_ratio,
+                    screen.near_plane() as f64 * ratio_threshold / screen_ratio,
+                    -screen.near_plane() as f64 * ratio_threshold,
+                    screen.near_plane() as f64 * ratio_threshold,
                     0.1,
                     screen.far_plane() as f64,
                 );
